alubm

+++ application/api_resources/song.py
@@ -14,6 +14,7 @@ songFields = {
     'title': fields.String,
     'views': fields.String,
     'genre' : fields.String,
+    'description' : fields.String,
 }
 class songPlaylistResourse(Resource):
     @login_required 
@@ -23,7 +24,7 @@ class songPlaylistResourse(Resource):
         song = Song.query.get_or_404(song_id)
 
         songplaylist = SongPlaylist.query.filter(SongPlaylist.playlist_id == playlist_id, SongPlaylist.song_id == song_id).first()
-        if songplaylist:
+        if songplaylist or playlist.user_id != current_user.id:
             return 'Failed', 406
 
         playlist.songs.append(song)
@@ -34,8 +35,9 @@ class songPlaylistResourse(Resource):
 
     @login_required 
     def delete(self, playlist_id, song_id):
+        playlist = Playlist.query.get_or_404(playlist_id)
         songplaylist = SongPlaylist.query.filter(SongPlaylist.playlist_id == playlist_id, SongPlaylist.song_id == song_id).first()
-        if songplaylist:
+        if songplaylist and playlist.user_id == current_user.id:
             db.session.delete(songplaylist)
             db.session.commit()
             return 'Success', 202
@@ -57,6 +59,29 @@ class albumSongSearchResource(Resource):
     def get(self):
         args = request.args
         query = db.select(Song, func.avg(SongLikes.rating).label('rating')).join(SongLikes , SongLikes.song_id == Song.id,isouter=True).group_by(Song.id).where(Song.creator_id == current_user.id)
+
+        empty = ['', None, ' ']
+        if args.get('title') not in empty:
+            query = query.where(Song.title.ilike('%'+args['title']+'%'))
+        if args.get('genre') not in empty and args.get('genre')!= 'all':
+            query = query.where(Song.genre == args['genre'])
+        if args.get('language') not in empty and args.get('genre')!= 'all':
+            query = query.where(Song.genre == args['language'])
+
+        res = db.session.execute(query).fetchall()
+        if res[0] == (None, None):
+            return 'Not found', 400
+        an = [{'song': r[0], 'rating':r[1]} for r in res]
+        return an
+    
+class songSearchResource(Resource):
+    parser = reqparse.RequestParser()
+    parser.add_argument('title')
+    @marshal_with(songQueryFields)
+    @login_required 
+    def get(self):
+        args = request.args
+        query = db.select(Song, func.avg(SongLikes.rating).label('rating')).join(SongLikes , SongLikes.song_id == Song.id,isouter=True).group_by(Song.id).limit(100)
 
         empty = ['', None, ' ']
         if args.get('title') not in empty:
